+++ CHANGES.rst
@@ -57,14 +57,14 @@ Specviz2d
 
 Other Changes and Additions
 ---------------------------
-- Temporarily disabled Spline Stretch registration due to incompatibility
-  with the current `glue` version. [#2585]
 
 3.8.1 (unreleased)
 ==================
 
 Bug Fixes
 ---------
+- Temporarily disabled Spline Stretch registration due to incompatibility
+  with the current `glue` version. [#2585]
 
 Cubeviz
 ^^^^^^^