+++ CHANGES.rst
@@ -4,6 +4,8 @@
 Bug Fixes
 ---------
 
+- Compatibility with glue-core 1.17. [#2591]
+
 Cubeviz
 ^^^^^^^