+++ CHANGES.rst
@@ -73,6 +73,8 @@ Cubeviz
 Imviz
 ^^^^^
 
+- Line Profile (XY) plugin no longer malfunctions when image contains NaN values. [#2594]
+
 Mosviz
 ^^^^^^