+++ CHANGES.rst
@@ -4,7 +4,10 @@
 Bug Fixes
 ---------
 
-- Compatibility with glue-core 1.17. [#2591]
+- Compatibility with glue-core 1.17. [#2591, #2595]
+
+- Fix image layer visibility toggle in plot options. [#2595]
+
 
 Cubeviz
 ^^^^^^^