+++ CHANGES.rst
@@ -15,6 +15,8 @@ Cubeviz
 Imviz
 ^^^^^
 
+- Line Profile (XY) plugin no longer malfunctions when image contains NaN values. [#2594]
+
 Mosviz
 ^^^^^^