+++ CHANGES.rst
@@ -9,6 +9,8 @@ Bug Fixes
 - Fix image layer visibility toggle in plot options. [#2595]
 
 
+- Fixes viewer toolbar items losing ability to bring up right-click menu. [#2605]
+
 Cubeviz
 ^^^^^^^